rewrite updateWatchlist cache flow
